// part9-typescript/patientor-frontend-TS/src/types.ts
export interface Diagnosis {
  code: string;
  name: string;
  latin?: string;
}

export enum Gender {
  Male = "male",
  Female = "female",
  Other = "other"
}

export interface BaseEntry {
  id: number;
  description: string;
  date: string;
  specialist: string;
  diagnosisCodes?: Array<Diagnosis['code']>;
}
export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3
}
interface Discharge {
  date: string,
  criteria: string
}
interface  SickLeave {
  startDate: string,
  endDate: string,
}
export interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}
interface HospitalEntry extends BaseEntry{
  type: 'Hospital',
  discharge : Discharge
}
interface OccupationalHealthcareEntry extends BaseEntry{
  type : 'OccupationalHealthcare',
  employerName: string,
  sickLeave?: SickLeave
}
export type newHealthCheckEntry  = Omit<HealthCheckEntry,'id'>
export type EntryTypes =  
| HospitalEntry['type']
| OccupationalHealthcareEntry['type']
| HealthCheckEntry['type'];
export type Entry =
  | HospitalEntry
  | OccupationalHealthcareEntry
  | HealthCheckEntry;

export interface Patient {
  id: number;
  name: string;
  ssn?: string;
  occupation: string;
  gender: Gender;
  dateOfBirth?: string;
  entries: Entry[]
}